backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
openai==1.3.0
anthropic==0.7.8
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1